changing file structure and improving makefile
