--- lemmo.emoji
@@ -1,3 +1,0 @@
-Use me when pushing please:::
-
-🍋🍋🍋🍋🍋🍋🍋🍋🍋🍋🍋🍋🍋🍋
