+++ src/algo_path.c
