+++ src/algo_path_more.c
