+++ src/anthill_hatchery.c
