+++ src/bfs_main.c
@@ -1,4 +1,4 @@
-#include "./includes/lem_in.h"
+#include <lem_in.h>
 
 bool	direct_link(t_link *link)
 {
