+++ src/display_ants.c
