+++ src/display_rooms.c
