+++ src/init_link.c
