+++ src/number_of_rooms.c
