+++ src/reset_rooms.c
