+++ src/room_functions.c
