+++ src/stdin_start_end.c
