+++ src/stdindata_init.c
