+++ src/validate_paths.c
