+++ src/verify_links.c
