+++ src/verify_room.c
