+++ src/verify_start_and_end.c
