--- testing.txt
@@ -1,38 +1,0 @@
-t_pathlist	*current = anthill->paths;
-while (current)
-{
-	printf("PATH LENGTH : %d\n", current->length);
-	current = current->next;
-}
-
-t_room		*cursor;
-cursor = anthill->linear;
-printf("ANTS : %d\n", anthill->nb_ants);
-printf("ROOM COUNT : %d\n\n", anthill->room_count);
-while (cursor != NULL)
-{
-	// if (cursor->is_end == 1 || cursor->is_start == 1)
-	// {
-	printf("NAME : %s\n", cursor->name);
-	printf("X POS : %f\n", cursor->x);
-	printf("Y POS : %f\n", cursor->y);
-	printf("IS_END : %d\n", cursor->is_end);
-	printf("IS_START : %d\n", cursor->is_start);
-	printf("INDEX : %d\n", cursor->index);
-	printf("LINK COUNT : %d\n", cursor->link_count);
-	printf("LEVEL : %d\n", cursor->level);
-	printf("\n\n");
-	// }
-	cursor = cursor->next;
-}
-while (1);
-
-t_link	*current;
-current = anthill->connectors;
-while (current)
-{
-	ft_putendl(current->from->name);
-	ft_putendl(current->to->name);
-	ft_putchar('\n');
-	current = current->next;
-}
